+++ src/main/java/daos/AgenciaLocalDao.java
@@ -72,12 +72,12 @@ public class AgenciaLocalDao {
         return lista;
     }
 
-    public static List<AgenciaLocal> getLocaisByAgenciaId(int idAgencia) {
+    public static List<AgenciaLocal> getLocaisByAgenciaId(Long idAgencia) {
         List<AgenciaLocal> lista = new ArrayList<>();
         try (Connection con = ConexaoDB.getConexao()) {
             String sql = "SELECT * FROM tb_agencia_local WHERE id_agencia = ?";
             PreparedStatement stm = con.prepareStatement(sql);
-            stm.setInt(1, idAgencia);
+            stm.setLong(1, idAgencia);
             ResultSet rs = stm.executeQuery();
 
             while (rs.next()) {
